memstore: don't encode/decode 'size_t'
there is no encode/decode funtions for type 'size_t'. We get following
error on OSX:

os/PageSet.h:207:5: error: call to 'encode' is ambiguous
os/MemStore.h:144:7: error: call to 'encode' is ambiguous

+++ src/os/MemStore.h
@@ -123,7 +123,7 @@ public:
 
   struct PageSetObject : public Object {
     PageSet data;
-    size_t data_len;
+    uint64_t data_len;
 
     // use a thread-local vector for the pages returned by PageSet, so we
     // can avoid allocations in read/write()
